requirements.txt
django 
djangorestframework
environs
pandas
psycopg2